refactor: rename screen as page
