+++ tig/lib/presentation/page/option/provider/option_notifier_provider.dart
@@ -1,0 +1,7 @@
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:tig/presentation/page/option/provider/state/option_notifier.dart';
+import 'package:tig/presentation/page/option/provider/state/option_state.dart';
+
+final optionNotifierProvider =
+    AutoDisposeStateNotifierProvider<OptionNotifier, OptionState>(
+        (ref) => OptionNotifier());
