+++ tig/lib/presentation/page/tag/provider/tag_notifier_provider.dart
@@ -1,0 +1,7 @@
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:tig/presentation/page/tag/provider/state/tag_notifier.dart';
+import 'package:tig/presentation/page/tag/provider/state/tag_state.dart';
+
+final tagNotifierProvider =
+    AutoDisposeStateNotifierProvider<TagNotifier, TagState>(
+        (ref) => TagNotifier());
