--- tig/lib/presentation/screens/auth/provider/auth_notifier_provider.dart
@@ -1,7 +1,0 @@
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:tig/presentation/screens/auth/provider/state/auth_notifier.dart';
-import 'package:tig/presentation/screens/auth/provider/state/auth_state.dart';
-
-final authNotifierProvider =
-    AutoDisposeStateNotifierProvider<AuthNotifier, AuthState>(
-        (ref) => AuthNotifier());
