--- tig/lib/presentation/screens/menu/provider/menu_notifier_provider.dart
@@ -1,7 +1,0 @@
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:tig/presentation/screens/menu/provider/state/menu_notifier.dart';
-import 'package:tig/presentation/screens/menu/provider/state/menu_state.dart';
-
-final menuNotifierProvider =
-    AutoDisposeStateNotifierProvider<MenuNotifier, MenuState>(
-        (ref) => MenuNotifier());
